requirements.txt
fastapi
uvicorn[standard]
openai
pandas
langdetect
python-dotenv
num2words
gtts
jinja2
